+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/MpaRowMapper2.java
@@ -10,7 +10,7 @@ public class MpaRowMapper2 implements RowMapper<Mpa> {
     @Override
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
-        mpa.setId(rs.getInt("mpa_rating_id"));
+        mpa.setId(rs.getLong("mpa_rating_id")); // Исправлено с getInt на getLong
         mpa.setName(rs.getString("mpa_name"));
         mpa.setDescription(rs.getString("mpa_description"));
         return mpa;
